+++ Changelog.md
@@ -1,5 +1,7 @@
 # Unreleased
-* `Configurate::Proxy` returns its own singleton class if the target does not support creating one
+
+* `Configurate::Proxy` returns its own singleton class if the target does not support creating one.
+* Extract `Configurate::Provider::StringHash` as a new base class for hash based providers.
 
 # 0.3.1